chore(mcp): consolidate ui app tests into co-located suite
Fold the unit/contract/render test layers into a single test file
co-located with the components, covering charts + insights logic only:
visualization inference (type guards, precedence, query-kind fallbacks)
and the chart/format utils, all driven by captured API fixtures shared
with the frontend's MSW mocks. Drop the tool-result contract tests, the
React render harness, and the actors-transform test, and revert the
config they needed.

Generated-By: PostHog Code
Task-Id: 8a135279-2519-48dd-ba25-2d52eaa39b0d
