+++ services/mcp/src/ui-apps/components/insightFixtures.ts
@@ -5,20 +5,20 @@
 // guards and visualizers inspect. `queryPayload` re-wraps a fixture's series as
 // the `{ query, results }` payload the MCP UI apps receive.
 
-import dataVisualizationHogQLJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/dataVisualizationHogQL.json'
-import funnelHistoricalTrendsJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/funnelHistoricalTrends.json'
-import funnelTimeToConvertJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/funnelTimeToConvert.json'
-import funnelTopToBottomBreakdownJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/funnelTopToBottomBreakdown.json'
-import funnelTopToBottomJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/funnelTopToBottom.json'
-import lifecycleJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/lifecycle.json'
-import retentionJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/retention.json'
-import stickinessJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/stickiness.json'
-import trendsLineBreakdownJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsLineBreakdown.json'
-import trendsLineJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsLine.json'
-import trendsNumberJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsNumber.json'
-import trendsPieJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsPie.json'
-import trendsWorldMapJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsWorldMap.json'
-import userPathsJson from '../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/userPaths.json'
+import dataVisualizationHogQLJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/dataVisualizationHogQL.json'
+import funnelHistoricalTrendsJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/funnelHistoricalTrends.json'
+import funnelTimeToConvertJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/funnelTimeToConvert.json'
+import funnelTopToBottomJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/funnelTopToBottom.json'
+import funnelTopToBottomBreakdownJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/funnelTopToBottomBreakdown.json'
+import lifecycleJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/lifecycle.json'
+import retentionJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/retention.json'
+import stickinessJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/stickiness.json'
+import trendsLineJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsLine.json'
+import trendsLineBreakdownJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsLineBreakdown.json'
+import trendsNumberJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsNumber.json'
+import trendsPieJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsPie.json'
+import trendsWorldMapJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/trendsWorldMap.json'
+import userPathsJson from '../../../../../frontend/src/mocks/fixtures/api/projects/team_id/insights/userPaths.json'
 
 export const insightResults = {
     trendsLine: trendsLineJson.result,
